starting to convert to redux
