+++ scripts/hello-node/game.js
@@ -1,21 +1,8 @@
 const R = require('ramda')
 
-const isEven = x => x % 2 === 0
 const BOARD_SIZE = 3
 
-const createEmptyBoard = () =>
-  R.range(0, 9)
 
-const createBoardWithMarkers = () =>
-  R.range(0, 9).map(R.ifElse(isEven, R.always('a'), R.always('b')))
-
-let createState = () => (
-  {
-    cursorPos: [0, 0],
-    // board: createEmptyBoard()
-    board: createBoardWithMarkers()
-  }
-);
 
 const indexForGrid = ({row, col}) => 3*row + col;
 
@@ -23,9 +10,7 @@ const indexForGrid = ({row, col}) => 3*row + col;
 const playerAtGrid = (state, {row, col}) =>
   state[indexForGrid({row, col})]
 
-function checkerboardBgColor (row, col) {
-  return isEven(row + col) ? 'blue' : 'darkBlue'
-}
+
 
 let clampWithinBoard = R.clamp(0, BOARD_SIZE - 1)
 const moveWithinBoard = (xfn, yfn) =>
@@ -36,8 +21,6 @@ const moveWithinBoard = (xfn, yfn) =>
 
 module.exports = {
   BOARD_SIZE,
-  checkerboardBgColor,
   moveWithinBoard,
   playerAtGrid,
-  createState
 }
