+++ scripts/hello-node/main.js
@@ -6,10 +6,14 @@ const ScreenBuffer = termkit.ScreenBuffer;
 const Redux = require('redux');
 const { createStore, applyMiddleware, combineReducers } = Redux;
 const reducer = require('./game-reducer').reducer;
-
+const Rendering = require('./rendering');
+const {
+  drawSquare,
+  drawCursor,
+  drawGrid
+} = Rendering;
 
 const Game = require('./game');
-let state = Game.createState();
 let viewport;
 
 const store = createStore(reducer);
@@ -81,59 +85,8 @@ function terminate () {
   }, 100);
 }
 
-function drawSquare (screenbuffer, {row, col, left, top, w, h, bgColor, char = ' '}) {
-  R.forEach((x) => {
-    R.forEach((y) => {
-      screenbuffer.put({
-        x: left + x,
-        y: top + y,
-        attr: {bgColor}
-      }, char);
-    }, R.range(1, h + 1));
-  }, R.range(1, w + 1));
-}
-
-
-function drawCursor(screenbuffer, {left, top, w, h, bgColor}) {
-  drawSquare(screenbuffer, {left, top, w, h, bgColor, char: '+'});
-}
-
-function drawPlayerMarker(whichPlayer, {left, top, w, h, bgColor} ) {
-  drawSquare(viewport, {left, top, w, h, bgColor, char: whichPlayer});
-}
-
-
-function drawGrid (screenbuffer) {
-  const SQUARE_WIDTH = 10;
-  const SQUARE_HEIGHT = 5;
-  const w = SQUARE_WIDTH;
-  const h = SQUARE_HEIGHT;
-
-  R.forEach((row) => {
-    R.forEach((col) => {
-      const bgColor = Game.checkerboardBgColor(row, col);
-
-      const left = (col * (SQUARE_WIDTH));
-      const top = (row * SQUARE_HEIGHT) + 1;
-      drawSquare(screenbuffer, {
-        left, top, w, h,
-        bgColor
-      });
-
-      if (R.equals(state.cursorPos,[col, row])) {
-        drawCursor(screenbuffer, {left, top, w, h, bgColor});
-      }
-
-      drawPlayerMarker(
-        Game.playerAtGrid(state.board, {row,col}),
-        {left, top, w: w/2, h: h/2}
-        );
-
-    }, R.range(0, Game.BOARD_SIZE));
-  }, R.range(0, Game.BOARD_SIZE));
-}
 
 init(() => {
-  drawGrid(viewport);
+  drawGrid(viewport, store);
   viewport.draw();
 })
